Fix crashing dialogs
The category select and locale select dialogs were crashing due to
`getDialog().getWindow().requestFeature(Window.FEATURE_NO_TITLE)` being
called after content was added to the dialog (according to the error
message). Moving this line from `onViewCreated` to `onCreateView` (which
is how all the other dialogs do it) seems to fix the problem.